% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_go.R
\docType{data}
\name{bp_go}
\alias{bp_go}
\title{bp_go data set includes all the go terms for Biological process. It is loaded in GOGOgadget package for GOGOscavenge to determine if the go terms is BP.}
\format{data.frame}
\usage{
data("bp_go")
}
\description{
bp_go data set includes all the go terms for Biological process. It is loaded in GOGOgadget package for GOGOscavenge to determine if the go terms is BP.
}
\references{
gene onthology consortium
}
\keyword{datasets}

