% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_go.R
\docType{data}
\name{cc_go}
\alias{cc_go}
\title{cc_go data set includes all the go terms for cellular component. It is loaded in GOGOgadget package for GOGOscavenge to determine if the go terms is CC.}
\format{data.frame}
\usage{
data("cc_go")
}
\description{
cc_go data set includes all the go terms for cellular component. It is loaded in GOGOgadget package for GOGOscavenge to determine if the go terms is CC.
}
\references{
gene onthology consortium
}
\keyword{datasets}

