% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogodictionary.R
\name{gogodictionary}
\alias{gogodictionary}
\title{GOGOGADGET DATA DICTIONARY
creates  print class and levels and whatnot}
\usage{
gogodictionary(dataset, NAcode, printalllevels = F)
}
\arguments{
\item{dataset}{A data frame name.}

\item{NAcode}{A quoted string ie "NA" or whatever your NA code is.}

\item{printalllevels}{TRUE OR FALSE print all levels of your variables}
}
\value{
names and column number of variables or matches.
}
\description{
GOGOGADGET DATA DICTIONARY
creates  print class and levels and whatnot
}
\examples{
gogodictionary(mtcars, "NA", printalllevels=F)

}
