% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogofindvar.R
\name{gogofindvar}
\alias{gogofindvar}
\title{Find a variable in a data set or partial matches of names.}
\usage{
gogofindvar(var, dataset)
}
\arguments{
\item{var}{A quoted string ie variable name or part of the name.}

\item{dataset}{A data frame name.}
}
\value{
names and column number of variables or matches.
}
\description{
Find a variable in a data set or partial matches of names.
}
\examples{
gogofindvar("mgp", mtcars)
gogofindvar("gear", mtcars)

}
