% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogogadget.R
\docType{package}
\name{gogogadget-package}
\alias{gogogadget}
\alias{gogogadget-package}
\title{GOGOGADET: A package of cool functions you wish you made.}
\description{
Right now it only has 4 functions but will contain more.
}
\section{Gogogogadet functions}{

The gogogadet functions, gogofindvar, gogogadetexport, gogostr, gogooperdir
}

\keyword{internal}
