% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogoexport.R
\name{gogogadgetexport}
\alias{gogogadgetexport}
\title{Write CSV file to specified location No row names}
\usage{
gogogadgetexport(dfr, loc = "F:\\\\A=Rexports\\\\")
}
\arguments{
\item{dfr}{A data frame name.}

\item{loc}{Output folder path. Default is "F:\\A=Rexports\\" for Moffitt employees.}
}
\value{
Outputs a csv file
}
\description{
Write CSV file to specified location No row names
}
\examples{
gogogadgetexport(mtcars)

}
