% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOGOsummary.R
\name{GOGOsummary}
\alias{GOGOsummary}
\title{GOGOsummary function returns pie charts showing the percentage of GO terms according to the input GO types.}
\usage{
GOGOsummary(scavenger, gotype, ...)
}
\description{
2 arguments: scavenger dataset processed by GOGOscavenger function and choose a GO type from "MF", "BP", "CC", "PC"
}
\examples{
\dontrun{
GOGOsummary(scav, "MF")
}

}
\author{
Xiaoping Li
}

