% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_go.R
\docType{data}
\name{mf_go}
\alias{mf_go}
\title{mf_go data set includes all the go terms for Molecular function. It is loaded in GOGOgadget package for GOGOscavenge to determine if the go terms is MF.}
\format{data.frame}
\usage{
data("mf_go")
}
\description{
mf_go data set includes all the go terms for Molecular function. It is loaded in GOGOgadget package for GOGOscavenge to determine if the go terms is MF.
}
\references{
gene onthology consortium
}
\keyword{datasets}

