% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotIDMaps.R
\docType{data}
\name{UniprotIDMaps}
\alias{UniprotIDMaps}
\title{UniprotIDMaps converts gene ID to Uniprot identifier}
\format{A data frame}
\usage{
data("UniprotIDMaps")
}
\description{
UniprotIDMaps converts gene ID to Uniprot identifier
}
\keyword{datasets}

