% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogodictionary.R
\name{gogodictionary}
\alias{gogodictionary}
\title{gogodictionary}
\usage{
gogodictionary(dataset, NAcode, printalllevels = F)
}
\arguments{
\item{dataset}{a data frame or maybe a tibble these days whatever your data identies as, right?}

\item{NAcode}{NA code only think for now}

\item{printalllevels}{T/F should all levels of a factor be printed, what out!}
}
\value{
a data frame ?
}
\description{
gogodictionary
}
