% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogofindvar.R
\name{gogofindvar}
\alias{gogofindvar}
\title{gogofindvar}
\usage{
gogofindvar(var, dataset)
}
\arguments{
\item{var}{a character string to find in the names of the data frame or tibble maybe. should be in quotes}

\item{dataset}{a data frame or maybe a tibble in this day and age whatever your data identifies as, right?}
}
\value{
a data frame
}
\description{
gogofindvar
}
