% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogogadgetexport.R
\name{gogogadgetexport}
\alias{gogogadgetexport}
\title{gogogadgetexport}
\usage{
gogogadgetexport(dfr, loc = "F:\\\\A=Rexports\\\\")
}
\arguments{
\item{dfr}{what to save; saves a .csv file in `F: A=Rexports ``}

\item{loc}{where to save}
}
\value{
a csv file
}
\description{
gogogadgetexport
}
