% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gogoopendir.R
\name{gogoopendir}
\alias{gogoopendir}
\title{gogoopendir
Opens your working directory!}
\usage{
gogoopendir(dir = getwd())
}
\arguments{
\item{dir}{getwd()}
}
\value{
opens a directory window
}
\description{
gogoopendir
Opens your working directory!
}
