% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOGOpick.R
\name{GOGOpick}
\alias{GOGOpick}
\title{GOGOpick automatically adds goTerms and other elements to the edgeR or DESeq2 processed file.}
\usage{
GOGOpick(dataset, database, write = TRUE, ...)
}
\description{
GOGOpick takes 4 arguments: "dataset" contains differentially expressed genes from edgeR and DESeq2. "database" is the Ontology database file you created use "DBconstruct" function. write" defualt is TRUE would write results in a csv file in the current directory. Use ... to specify the name of the csv file.
}
\examples{
\dontrun{GOGOpick("DE_drad_24hr_adjst.csv", [dra]OntologyDataBase, write = T, "Up and Down regulated genes by using DESeq2.csv")}

}
\author{
Xiaoping Li
}

