% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotIDMaps.R
\docType{data}
\name{UniprotIDMaps}
\alias{UniprotIDMaps}
\title{UniprotIDMaps}
\format{a data.frame}
\usage{
data("UniprotIDMaps")
}
\description{
conversion of gene ID and Uniprot identifier
}
\references{
uniprotKB
}
\keyword{datasets}

