% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotIDMaps.R
\docType{data}
\name{UniprotIDMaps}
\alias{UniprotIDMaps}
\title{UniprotIDMaps data}
\format{data.frame}
\usage{
data("UniprotIDMaps")
}
\description{
dataset for mapping refID (gene ID) to Uniprot ID, including symbols. Also loaded for GOGOscavenger function.
}
\references{
http://www.uniprot.org/uploadlists/
}
\keyword{datasets}

