% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniprotIDMaps.R
\docType{data}
\name{UniprotIDMaps}
\alias{UniprotIDMaps}
\title{UniprotIDMaps data. dataset for mapping refID or gene ID to Uniprot ID, including symbols. Also loaded for GOGOscavenger function.}
\format{data.frame}
\usage{
data("UniprotIDMaps")
}
\description{
UniprotIDMaps data. dataset for mapping refID or gene ID to Uniprot ID, including symbols. Also loaded for GOGOscavenger function.
}
\references{
Uniprot
}
\keyword{datasets}

